+++ inc/hooks.php
@@ -29,7 +29,6 @@ add_action( 'widgets_init', __NAMESPACE__ . '\widgets_init' );
  * Scripts and styles associated hooks
  */
 require get_theme_file_path( 'inc/hooks/scripts-styles.php' );
-add_action( 'wp_enqueue_scripts', __NAMESPACE__ . '\enqueue_polyfills' );
 add_action( 'wp_enqueue_scripts', __NAMESPACE__ . '\enqueue_theme_scripts' );
 
 // NB! If you use ajax functionality in Gravity Forms, remove this line